part2

+++ libft/part1/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: seongwch <seongwch@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 15:08:49 by seongwch          #+#    #+#             */
-/*   Updated: 2021/11/11 15:27:24 by seongwch         ###   ########.fr       */
+/*   Updated: 2021/11/12 16:31:21 by seongwch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t num)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < num && s1[i] != '\0' && s2[i] != '\0')
